requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==1.4.50
WTForms==3.0.1
reportlab==4.0.8
gunicorn==21.2.0
python-dotenv==1.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
flask-dance==7.0.1
psycopg2-binary==2.9.9
